+++ docs/release-notes/change-log.md
@@ -9,6 +9,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* added store with update policy `set_sum` which allows the store to either sum a numerical value, or set it to a new value
+
 ## v1.6.2
 
 * execout: preload only one file instead of two, log if undeleted caches found